Create a single-stream Observable from an Iterable, instead of wrapping it
If the response body is an Iterable, it makes sense to return an Observable<R>, instead of
Observable<Iterable<R>>. This will make APIs stick better to the core idea of observables being
streams by default. It will also help avoid having to use flatMap, in order to convert to
Observable<R>

TODO: Consider applying the same logic for all adapters (incl. Result and Response)